[patch] Merge comparison scripts into main examples
