+++ tests/README.md
@@ -1,2 +1,6 @@
-These tests run the Able interpreter against example files and verify the expected output.
-Use `python3 run_tests.py` to execute the suite.
+The test suite compiles the interpreter and runs Able scripts found in the
+`examples/` directory. Test modules live under `tests/integration` and use a
+shared helper for invoking the interpreter.
+
+Run `python3 run_tests.py` from the repository root to build the executable and
+execute all tests.
