+++ tests/integration/helpers.py
@@ -1,0 +1,16 @@
+import subprocess
+from pathlib import Path
+import unittest
+
+EXE = Path('build/able_exe')
+
+class AbleTestCase(unittest.TestCase):
+    @classmethod
+    def setUpClass(cls):
+        subprocess.run(['make'], check=True)
+        if not EXE.exists():
+            raise RuntimeError('Executable not built')
+
+    def run_script(self, path: str) -> str:
+        result = subprocess.run([str(EXE), path], check=True, capture_output=True, text=True)
+        return result.stdout
